Fix oveeriding request options parameters in requestGatewayHelper
The arguments 'isJson' & isBinary are optional in requestGatewayHelper. When not given, they should have a default value.

Example with 'isJson': The previous implementation was setting the request option 'json' field to true by default before checking
wether isJson was defined (and if it was defined, what was its state).

The previous implementation tried to override the request option json field with Object.assign(). But, this is not what Object.assign()
is doing (or should be used for) => in case of overriding an already existing property, it is using the value of the last argument having it.

In our case, it was the initial 'options' variable which was provided last, hence we were keeping the default value of json => it was impossible to override
the json request property.

The same was happening for others request option properties. Now, it is fixed.