Add a shell script to generate package-lock.json for exs
The examples are tested with an `npm link` during the development of a
new version of the SDK which is bypassing the `package-lock.json`
generation.

When a new version of the SDK is released on npm repository, we need to
generate a new `package-lock.json` for each example before committing
those on GitHub so that they can be tagged with the new release.