Update compartment_id type
The compartment_id in the activities or events should be a `string` instead of a number.
I've updated the interface of api/datamart repository to reflect that.

Note that it should also be a string within the UserIdentifierInterface but as of today the back is using it as a `number`.
I've set a reminder in the code so that when the back is updated we'll also update the interface in the plugin sdk