Fix Handlebars types conflict
TLDR: Using handlebars 4.1.x and typedoc is not possible at the moment.
Hence, we stay on handlebars 4.0.x until next typedoc release.

It appears that:
- Handlebars is now directly releasing its own typescript bindings since
the 4.1.x version
- the @types/handlebars package is conflicting with the bindings that
comes with the last version of Handlebars. So it can't be used with the 4.1.x
releases of Handlebars
- typedoc is still using the 4.0.x version of handlebars + its
@types conterpart in its latest release (v0.14.2).
- As we depend on typedoc, we can't use the new
version of handlebars (4.1.x) in the SDK as the Typescript bindings from
the typedoc dependency and the bindings coming directly from Handlebars
are conflicting

So, while waiting for typedoc to upgrade to handlebars v4.1.x so that
we can remove all dependency on @types/handlebars in the SDKL, we need
to keep the SDK at the v4.0.x version of handlebars + its @types
conterpart to solve the conflict issue in the short term.

Such a beautiful world.