Add variability in the instance context refresh interval
In order to avoid having 'bursts' effects on the Gateway APIs calls that
are being done when the instance context is re-built, we're now adding
some variability in the refresh period (+10%).

We also updated the refresh period from 2min to 10min.

This should lower the Gateway API load and provide a better plugin
stability / performance.