Update dependencies and fix Handlebars type issue for good
In previous episodes:

We used to rely on:
- @types/handlebars
- handlebars
- typedoc

And everything was fine.

One day, the guys in handlebars decided to embed the types declaration
directly in their package *in a minor release*. And then there was a
conflict between
@types/handlebars and handlebars as we accept the minor releases
upgrades in package.json.

You might think that removing our dependency to @types/handlebars and
upgrade our dependency to the last release of handlebars would have fix
the conflict.

But then typedoc enter the dance -> their latest release still rely on
@types/handlebars. So we still had a transitive issue to
@types/handlebars and the conflict was still here.

Our plan was to:
- stick to the old version of handlebars that is not embedding the
type definition file
- wait for typedoc to release a new release without the dependency to
@types/handlebars (it's already commited on their master branch)

BUT, @types/handlebars released a minor version that is removing the
type definition.

And as we allowed the minor update, we're now without any definition
for Handlebars types.

But, that also mean that typedoc is no longer fetching the type
definition from @types/handlebars (they also accept minor releases of
this package). So no more conflicts from them!

So we just have to:
- Updgrade handlebars to the 4.x.x releases that embed the types
definition
- Remove @types/handlebars dependency
- Grab a beer to celebrate!

This commit is also doing an update of all packages.