FIX: Refuse to process calls before the initialization
With the lifecycle of the plugin, we should not get any request before
getting a `/v1/init` call and changing the health check response.
However, in some edge cases, we can get a race condition where a small
number of requests can still be sent. We don't have any credentials so
we can't prepare the context and can't properly handle the request.

However, if we don't stop ASAP the request we will save in the cache a
failed promise which will fail *all* requests for 10 minutes (for a
given instance id). Now, we still loose the extra early call(s) but
start processing correctly as soon as the initialization kicks in.