MICS-11819 Migrate to the new way to get credentials
The platform now gives credentials (worker_id/authentication_token) via
the environment. The `CREDENTIAL_UPDATE_FROM_*` mechanism can also be
removed as all processes will get the same env / the same credentials,
and they don't change anymore (for the lifetime of the container).
While not perfect, overriding the env in the tests code avoid a
dependency to a wrapper script.
This is a breaking change: the api changes, all tests must be fixed.